+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.reservation.ReservationResponses;
+import roomescape.dto.response.ReservationResponses;
 import roomescape.infrastructure.LoginUserId;
 import roomescape.service.AdminReservationService;
