+++ src/main/java/roomescape/controller/ReservationController.java
@@ -12,13 +12,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.dto.reservation.CancelReservationCommand;
-import roomescape.dto.reservation.CreateReservationCommand;
-import roomescape.dto.reservation.CreateReservationRequest;
-import roomescape.dto.reservation.ReservationResponse;
-import roomescape.dto.reservation.ReservationWithStatusResponses;
-import roomescape.dto.reservation.UpdateReservationCommand;
-import roomescape.dto.reservation.UpdateReservationRequest;
+import roomescape.dto.command.CancelReservationCommand;
+import roomescape.dto.command.CreateReservationCommand;
+import roomescape.dto.request.CreateReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.ReservationWithStatusResponses;
+import roomescape.dto.command.UpdateReservationCommand;
+import roomescape.dto.request.UpdateReservationRequest;
 import roomescape.infrastructure.LoginUserId;
 import roomescape.service.ReservationService;
