+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
-import roomescape.dto.reservationtime.ReservationTimeResponses;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.dto.response.ReservationTimeResponses;
 import roomescape.service.ReservationTimeService;
 
 @Validated
