--- src/main/java/roomescape/dto/auth/LoginRequest.java
@@ -1,9 +1,0 @@
-package roomescape.dto.auth;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @NotBlank String username,
-        @NotBlank String password
-) {
-}
