--- src/main/java/roomescape/dto/auth/LoginResponse.java
@@ -1,6 +1,0 @@
-package roomescape.dto.auth;
-
-public record LoginResponse(
-        String accessToken
-) {
-}
