--- src/main/java/roomescape/dto/reservation/CancelReservationCommand.java
@@ -1,7 +1,0 @@
-package roomescape.dto.reservation;
-
-public record CancelReservationCommand(
-        Long reservationId,
-        Long userId
-) {
-}
