--- src/main/java/roomescape/dto/reservation/CreateReservationRequest.java
@@ -1,12 +1,0 @@
-package roomescape.dto.reservation;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-
-public record CreateReservationRequest(
-        @NotNull Long themeId,
-        @NotNull LocalDate date,
-        @NotNull Long timeId,
-        @NotNull Long storeId
-) {
-}
