--- src/main/java/roomescape/dto/reservationtime/CreateReservationTimeRequest.java
@@ -1,9 +1,0 @@
-package roomescape.dto.reservationtime;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalTime;
-
-public record CreateReservationTimeRequest(
-        @NotNull LocalTime startAt
-) {
-}
