--- src/main/java/roomescape/dto/reservationtime/ReservationTimeResponse.java
@@ -1,15 +1,0 @@
-package roomescape.dto.reservationtime;
-
-import roomescape.domain.ReservationTime;
-
-public record ReservationTimeResponse(
-        Long id,
-        String startAt
-) {
-    public static ReservationTimeResponse from(ReservationTime reservationTime) {
-        return new ReservationTimeResponse(
-                reservationTime.getId(),
-                reservationTime.getStartAt().toString()
-        );
-    }
-}
