--- src/main/java/roomescape/dto/reservationtime/ReservationTimeResponses.java
@@ -1,16 +1,0 @@
-package roomescape.dto.reservationtime;
-
-import java.util.List;
-import roomescape.domain.ReservationTime;
-
-public record ReservationTimeResponses(
-        List<ReservationTimeResponse> times,
-        boolean hasNext
-) {
-    public static ReservationTimeResponses of(List<ReservationTime> reservationTimes, boolean hasNext) {
-        return new ReservationTimeResponses(
-                reservationTimes.stream().map(ReservationTimeResponse::from).toList(),
-                hasNext
-        );
-    }
-}
