--- src/main/java/roomescape/dto/theme/CreateThemeRequest.java
@@ -1,10 +1,0 @@
-package roomescape.dto.theme;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record CreateThemeRequest(
-        @NotBlank String name,
-        @NotBlank String description,
-        @NotBlank String thumbnailImageUrl
-) {
-}
