--- src/main/java/roomescape/dto/theme/PopularThemeResponse.java
@@ -1,23 +1,0 @@
-package roomescape.dto.theme;
-
-import roomescape.domain.PopularTheme;
-import roomescape.domain.Theme;
-
-public record PopularThemeResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnailImageUrl,
-        Long reservedCount
-) {
-    public static PopularThemeResponse from(PopularTheme popularTheme) {
-        Theme theme = popularTheme.getTheme();
-        return new PopularThemeResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnailImageUrl(),
-                popularTheme.getReservationCount()
-        );
-    }
-}
