--- src/main/java/roomescape/dto/theme/PopularThemeResponses.java
@@ -1,14 +1,0 @@
-package roomescape.dto.theme;
-
-import java.util.List;
-import roomescape.domain.PopularTheme;
-
-public record PopularThemeResponses(
-        List<PopularThemeResponse> themes
-) {
-    public static PopularThemeResponses from(List<PopularTheme> popularThemes) {
-        return new PopularThemeResponses(
-                popularThemes.stream().map(PopularThemeResponse::from).toList()
-        );
-    }
-}
