--- src/main/java/roomescape/dto/theme/ThemeReservationTimeResponse.java
@@ -1,8 +1,0 @@
-package roomescape.dto.theme;
-
-public record ThemeReservationTimeResponse(
-        Long id,
-        String startAt,
-        boolean isReserved
-) {
-}
