--- src/main/java/roomescape/dto/theme/ThemeReservationTimeResponses.java
@@ -1,11 +1,0 @@
-package roomescape.dto.theme;
-
-import java.util.List;
-
-public record ThemeReservationTimeResponses(
-        List<ThemeReservationTimeResponse> times
-) {
-    public static ThemeReservationTimeResponses from(List<ThemeReservationTimeResponse> times) {
-        return new ThemeReservationTimeResponses(times);
-    }
-}
