--- src/main/java/roomescape/dto/theme/ThemeResponse.java
@@ -1,19 +1,0 @@
-package roomescape.dto.theme;
-
-import roomescape.domain.Theme;
-
-public record ThemeResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnailImageUrl
-) {
-    public static ThemeResponse from(Theme theme) {
-        return new ThemeResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnailImageUrl()
-        );
-    }
-}
