--- src/main/java/roomescape/dto/user/CreateUserRequest.java
@@ -1,10 +1,0 @@
-package roomescape.dto.user;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record CreateUserRequest(
-        @NotBlank String username,
-        @NotBlank String password,
-        @NotBlank String name
-) {
-}
