+++ src/main/java/roomescape/service/ThemeService.java
@@ -7,9 +7,9 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.PopularTheme;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.theme.CreateThemeRequest;
-import roomescape.dto.theme.ThemeReservationTimeResponse;
-import roomescape.dto.theme.ThemeResponses;
+import roomescape.dto.request.CreateThemeRequest;
+import roomescape.dto.response.ThemeReservationTimeResponse;
+import roomescape.dto.response.ThemeResponses;
 import roomescape.exception.ResourceNotFoundException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
