+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -13,9 +13,9 @@ import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.User;
-import roomescape.dto.theme.CreateThemeRequest;
-import roomescape.dto.theme.ThemeReservationTimeResponse;
-import roomescape.dto.theme.ThemeResponses;
+import roomescape.dto.request.CreateThemeRequest;
+import roomescape.dto.response.ThemeReservationTimeResponse;
+import roomescape.dto.response.ThemeResponses;
 import roomescape.fixture.Fixtures;
 import roomescape.repository.fake.FakeReservationRepository;
 import roomescape.repository.fake.FakeReservationTimeRepository;
